requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==19.3.0             # via pytest
bcrypt==3.1.7
certifi==2019.9.11        # via cloudinary, requests
cffi==1.13.2              # via bcrypt
chardet==3.0.4            # via requests
cloudinary==1.18.2
facebook-sdk==3.1.0
falcon==2.0.0
future==0.18.2            # via python-twitter
gunicorn==20.0.0
idna==2.8                 # via requests
importlib-metadata==0.23  # via pluggy, pytest
mock==3.0.5               # via cloudinary
more-itertools==7.2.0     # via pytest, zipp
oauthlib==3.1.0           # via requests-oauthlib
packaging==19.2           # via pytest
peewee==3.11.2
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.4
py==1.8.0                 # via pytest
pycparser==2.19           # via cffi
pyparsing==2.4.5          # via packaging
pytest==5.3.1
python-datauri==0.2.8
python-dotenv==0.10.3
python-http-client==3.2.1  # via sendgrid
python-twitter==3.5
requests-oauthlib==1.3.0  # via python-twitter
requests==2.22.0          # via facebook-sdk, python-twitter, requests-oauthlib
sendgrid==6.1.0
six==1.13.0               # via bcrypt, cloudinary, mock, packaging
urllib3==1.25.7           # via cloudinary, requests
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.1        # via gunicorn
